feat: ability to approve cash payments
