+++ server/controllers/formPayment.controller.js
@@ -5,7 +5,7 @@ require("dotenv").config();
 const { FormPayment, Form } = require("../models");
 const Response = require("./../mongoModels/response");
 const AccountController = require("./account.controller");
-const UserController = require("./user.controller");
+const createPayerUser = require("../utilityFunctions/createPayerUser");
 
 const FormPaymentController = function () {
   var connectResponseToFormPayment = async function (responseData, responseId) {
@@ -313,19 +313,8 @@ const FormPaymentController = function () {
           status: 404,
         };
       }
-      const user = getUserDataFromAnswers(formattedAnswers, groupId);
-
-      const updatedUserResponse =
-        await UserController.createUserViaFromResponse(user);
-
-      if (!updatedUserResponse.success) {
-        return updatedUserResponse;
-      }
-
-      const updatedUser = updatedUserResponse.data;
-      const paymentData = {
-        payer_id: updatedUser.id,
-      };
+      // sets up user data:
+      paymentData = await createPayerUser(formattedAnswers, groupId);
 
       await existingFormPayment.update(paymentData, { valudate: true });
 
@@ -338,32 +327,6 @@ const FormPaymentController = function () {
       console.error(error.stack);
       return { success: false, error: error.message, status: 500 };
     }
-  };
-
-  var getUserDataFromAnswers = function (formattedAnswers, groupId) {
-    const user = {
-      first_name: formattedAnswers.first_name?.short_text,
-      last_name: formattedAnswers.last_name?.short_text,
-      email: formattedAnswers.email?.email,
-      phone_number: formattedAnswers.phone_number?.phone_number,
-      address: formattedAnswers.address?.long_text,
-      date: formattedAnswers.date?.date,
-      photo: formattedAnswers.photo?.photo,
-      signature: formattedAnswers.signature?.short_text,
-      liability: formattedAnswers.liability?.liability,
-      team_id: formattedAnswers.player?.player?.team_id,
-      team_name: formattedAnswers.player?.player?.team_name,
-      league_name: formattedAnswers.player?.player?.league_name,
-      league_id: formattedAnswers.player?.player?.league_id,
-      season_name: formattedAnswers.player?.player?.season_name,
-      season_id: formattedAnswers.player?.player?.season_id,
-      division_name: formattedAnswers.player?.player?.division_name,
-      division_id: formattedAnswers.player?.player?.division_id,
-      payment_intent_id: formattedAnswers.payment?.paymentIntentId,
-      payment_amount: formattedAnswers.payment?.amount,
-      group_id: groupId,
-    };
-    return user;
   };
 
   var updateFormPaymentType = async function (req, res, next) {
